Workshop 02. Review01. Sensors data replaced to the Ws02SensorManager.
